Converted to Float

+++ Chapter_1/Exercise_1-4.c
@@ -4,8 +4,8 @@
 //C = (5/9)(F-32)
 
 main(){
-    int fahr, celsius;
-    int lower, upper, step;
+    float fahr, celsius;
+    float lower, upper, step;
 
     lower = 0; //lower limit of the scale
     upper = 150; //upper limit of the scale
@@ -15,8 +15,8 @@ main(){
     printf("Celsius to Fahrenheit \n");
 
     while (celsius <= upper){
-        fahr = 5 * (celsius-32)/9;
-        printf("%d\t%d\n", celsius, fahr);
+        fahr = 5.0 * (celsius-32.0)/9.0;
+        printf("%.2f\t%.2f\n", celsius, fahr);
         celsius = celsius + step;
     }
 }
